Allow specifying a custom attribute prefix for custom attributes that are exported to the USD data.
This allows setting `ayon+settings://maya/publish/ExtractMayaUsd/custom_attr_prefix` to e.g. `userProperties:` so that a Maya custom attribute `helloworld` that matches the `attr` or `attr_prefixes` instance attribute values (which defines which custom attribute to include) will be be named `userProperties:helloworld` in the resulting USD data.